+++ src/main/java/net/contargo/types/contactinfo/ContactInfoConsumer.java
@@ -22,4 +22,10 @@ public interface ContactInfoConsumer {
 
         // OK
     }
+
+
+    default void reset() {
+
+        // OK
+    }
 }