+++ src/main/java/net/contargo/types/contactinfo/normalization/PhoneNumberNormalizer.java
@@ -1,8 +1,9 @@
 package net.contargo.types.contactinfo.normalization;
 
 import net.contargo.types.Loggable;
-import net.contargo.types.telephony.formatting.PhoneNumberFormatter;
-import net.contargo.types.telephony.formatting.PhoneNumberFormattingException;
+import net.contargo.types.telephony.PhoneNumber;
+
+import org.apache.commons.lang.StringUtils;
 
 import java.util.Optional;
 
@@ -11,16 +12,10 @@ public class PhoneNumberNormalizer implements Loggable {
 
     public Optional<String> normalizeNumber(String phoneNumber) {
 
-        if (phoneNumber == null) {
+        if (StringUtils.isBlank(phoneNumber)) {
             return Optional.empty();
         }
 
-        try {
-            return Optional.of(new PhoneNumberFormatter().parseAndFormatToE164Format(phoneNumber));
-        } catch (PhoneNumberFormattingException e) {
-            logger().warn("Failed to parse and format number {}: {}", phoneNumber, e.getMessage());
-
-            return Optional.empty();
-        }
+        return new PhoneNumber(phoneNumber).getInternationalFormatOfPhoneNumber();
     }
 }