+++ src/main/java/net/contargo/types/telephony/formatting/PhoneNumberFormatter.java
@@ -2,17 +2,37 @@ package net.contargo.types.telephony.formatting;
 
 import com.google.i18n.phonenumbers.NumberParseException;
 import com.google.i18n.phonenumbers.PhoneNumberUtil;
+import com.google.i18n.phonenumbers.PhoneNumberUtil.PhoneNumberType;
 import com.google.i18n.phonenumbers.Phonenumber;
+
+import net.contargo.types.Loggable;
+
+import org.apache.commons.lang.StringUtils;
+
+import java.util.Arrays;
 
 
 /**
  * Handles the formatting of telephone numbers.
  *
  * @author  Robin Jayasinghe - jayasinghe@synyx.de
+ * @author  Julia Dasch - dasch@synyx.de
  */
-public class PhoneNumberFormatter {
+public class PhoneNumberFormatter implements Loggable {
+
+    private static PhoneNumberFormatter instance = null;
 
     private final PhoneNumberUtil phoneNumberUtil = PhoneNumberUtil.getInstance();
+
+    public static PhoneNumberFormatter getInstance() {
+
+        if (instance == null) {
+            return new PhoneNumberFormatter();
+        }
+
+        return instance;
+    }
+
 
     /**
      * Formats a given phoneNumber according to the rules defined in the E164 standard:
@@ -26,15 +46,9 @@ public class PhoneNumberFormatter {
      */
     public String parseAndFormatToE164Format(final String phoneNumber) throws PhoneNumberFormattingException {
 
-        final String phoneNumberWithoutWhitespace = phoneNumber.replaceAll("\\s+", "");
+        final Phonenumber.PhoneNumber parsedNumber = parsePhoneNumber(phoneNumber, trimPhoneNumber(phoneNumber));
 
-        final Phonenumber.PhoneNumber parsedNumber;
-
-        parsedNumber = parsePhoneNumber(phoneNumber, phoneNumberWithoutWhitespace);
-
-        final String formattedNumber = phoneNumberUtil.format(parsedNumber, PhoneNumberUtil.PhoneNumberFormat.E164);
-
-        return formattedNumber;
+        return phoneNumberUtil.format(parsedNumber, PhoneNumberUtil.PhoneNumberFormat.E164);
     }
 
 
@@ -57,23 +71,30 @@ public class PhoneNumberFormatter {
      */
     public String parseAndFormatToDIN5008(final String phoneNumber) throws PhoneNumberFormattingException {
 
-        // strip all whitespace from input
-        final String phoneNumberWithoutWhitespace = phoneNumber.replaceAll("\\s+", "");
+        // strip all whitespace and line breaks from input
+        final String trimPhoneNumber = trimPhoneNumber(phoneNumber);
 
         // pre-validation and formatting with libphonenumber
-        final Phonenumber.PhoneNumber parsedNumber = parsePhoneNumber(phoneNumber, phoneNumberWithoutWhitespace);
+        final Phonenumber.PhoneNumber parsedNumber = parsePhoneNumber(phoneNumber, trimPhoneNumber);
         final String formattedNumber = phoneNumberUtil.format(parsedNumber,
                 PhoneNumberUtil.PhoneNumberFormat.INTERNATIONAL);
 
         // extract primitives from libphonenumber parse result
         final String nationalNumber = String.valueOf(parsedNumber.getNationalNumber());
         final int countryCode = parsedNumber.getCountryCode();
-        final PhoneNumberUtil.PhoneNumberType phoneNumberType = PhoneNumberUtil.getInstance()
-                .getNumberType(parsedNumber);
+        final PhoneNumberType phoneNumberType = getPhoneNumberType(phoneNumber);
 
         // trigger the actual formatting
-        return formatDIN5008(countryCode, nationalNumber, formattedNumber,
-                phoneNumberType.equals(PhoneNumberUtil.PhoneNumberType.FIXED_LINE));
+        String formatDIN5008 = formatDIN5008(countryCode, nationalNumber, formattedNumber,
+                phoneNumberType.equals(PhoneNumberType.FIXED_LINE));
+
+        if (formatDIN5008.length() >= trimPhoneNumber.length()) {
+            return formatDIN5008;
+        } else {
+            throw new PhoneNumberFormattingException(String.format(
+                    "can not format number. the formatted phone number('%s') has less digits than the incoming phone number('%s')",
+                    formatDIN5008, phoneNumber));
+        }
     }
 
 
@@ -95,7 +116,59 @@ public class PhoneNumberFormatter {
         final String formattedBaseNumber = parseAndFormatToDIN5008(basePhoneNumber);
         final String sanitizedExtension = extension.replaceAll("\\s+", "").replaceAll("\\D+", "");
 
-        return String.format("%s-%s", formattedBaseNumber, sanitizedExtension);
+        return String.format("%s %s", formattedBaseNumber, sanitizedExtension);
+    }
+
+
+    /**
+     * Returns the region code of a number. It is needed to check the whole number to return the region/country code.
+     * It will not only check the country calling code. example for +1 242 = 'BS'
+     *
+     * @param  phoneNumber  the phoneNumber to get the region code/country code from
+     *
+     * @return  the country code from formatted number
+     *
+     * @throws  PhoneNumberFormattingException  if the phoneNumber cannot be parsed
+     */
+    public String getRegionCode(final String phoneNumber) throws PhoneNumberFormattingException {
+
+        Phonenumber.PhoneNumber parsedNumber = parsePhoneNumber(phoneNumber, trimPhoneNumber(phoneNumber));
+
+        return phoneNumberUtil.getRegionCodeForNumber(parsedNumber);
+    }
+
+
+    /**
+     * Returns the national significant number of a phone number.
+     *
+     * @param  phoneNumber  the phoneNumber to get the national significant number from
+     *
+     * @return  the national significant number
+     *
+     * @throws  PhoneNumberFormattingException  if the phoneNumber cannot be parsed
+     */
+    public String getNationalSignificantNumber(final String phoneNumber) throws PhoneNumberFormattingException {
+
+        Phonenumber.PhoneNumber parsedNumber = parsePhoneNumber(phoneNumber, trimPhoneNumber(phoneNumber));
+
+        return phoneNumberUtil.getNationalSignificantNumber(parsedNumber);
+    }
+
+
+    /**
+     * Returns the type of a number.
+     *
+     * @param  phoneNumber  the phoneNumber to get the type from
+     *
+     * @return  phone number type
+     *
+     * @throws  PhoneNumberFormattingException  if the phoneNumber cannot be parsed
+     */
+    public PhoneNumberType getPhoneNumberType(final String phoneNumber) throws PhoneNumberFormattingException {
+
+        Phonenumber.PhoneNumber parsedNumber = parsePhoneNumber(phoneNumber, trimPhoneNumber(phoneNumber));
+
+        return phoneNumberUtil.getNumberType(parsedNumber);
     }
 
 
@@ -117,71 +190,92 @@ public class PhoneNumberFormatter {
     private String formatDIN5008(final int countryCode, final String nationalNumber, final String preFormattedNumber,
         final boolean isFixedLine) throws PhoneNumberFormattingException {
 
-        final AreaCodeAndConnectionNumber areaCodeAndConnectionNumber;
+        final NationalNumber areaCodeAndConnectionNumber;
 
         if (isFixedLine) {
-            areaCodeAndConnectionNumber = getAreaCodeAndConnectionNumberFromFixedNumber(preFormattedNumber);
+            areaCodeAndConnectionNumber = getNationalNumberFromFixedNumber(preFormattedNumber);
         } else {
-            areaCodeAndConnectionNumber = getAreaCodeAndConnectionNumberFromMobileNumber(nationalNumber);
+            areaCodeAndConnectionNumber = getNationalNumberFromMobileNumber(nationalNumber);
         }
 
         return chunkAndFormatResult(countryCode, areaCodeAndConnectionNumber);
     }
 
 
-    private AreaCodeAndConnectionNumber getAreaCodeAndConnectionNumberFromMobileNumber(String nationalNumber) {
-
-        AreaCodeAndConnectionNumber areaCodeAndConnectionNumber;
+    private NationalNumber getNationalNumberFromMobileNumber(String nationalNumber) {
 
         if (nationalNumber.length() > 4) { // only for numbers longer than 4 digits
-            areaCodeAndConnectionNumber = new AreaCodeAndConnectionNumber(nationalNumber.substring(0, 3),
-                    nationalNumber.substring(3, nationalNumber.length()).replaceAll("\\D", ""));
-        } else { // for the shorter numbers
-            areaCodeAndConnectionNumber = new AreaCodeAndConnectionNumber("", nationalNumber);
-        }
 
-        return areaCodeAndConnectionNumber;
+            String connectionNumber = nationalNumber.substring(3).replaceAll("\\D", "");
+
+            return new NationalNumber(nationalNumber.substring(0, 3), connectionNumber);
+        } else { // for the shorter numbers
+            return new NationalNumber("", nationalNumber);
+        }
     }
 
 
-    private AreaCodeAndConnectionNumber getAreaCodeAndConnectionNumberFromFixedNumber(String preFormattedNumber)
+    private NationalNumber getNationalNumberFromFixedNumber(String preFormattedNumber)
         throws PhoneNumberFormattingException {
 
-        AreaCodeAndConnectionNumber areaCodeAndConnectionNumber;
+        // fist element should be the country code
         final String[] splittedPreformattedNumber = preFormattedNumber.split(" ");
 
-        if (splittedPreformattedNumber.length == 0) { // no elements -> must not happen in theory ;)
+        if (splittedPreformattedNumber.length <= 1) { // no or one elements -> must not happen in theory ;)
             throw new PhoneNumberFormattingException(String.format("Could not extract anything from input number: %s",
                     preFormattedNumber));
-        } else if (splittedPreformattedNumber.length == 1) { // one element -> take the whole number as connection
-            areaCodeAndConnectionNumber = new AreaCodeAndConnectionNumber("", preFormattedNumber);
+        } else if (splittedPreformattedNumber.length <= 2) { // two elements -> take the second element as connection
+            return new NationalNumber("", splittedPreformattedNumber[1].replaceAll("\\D+", ""));
         } else { // more than one element -> take the second as area code and the third as connection number (first is country code)
-            areaCodeAndConnectionNumber = new AreaCodeAndConnectionNumber(preFormattedNumber.split(" ")[1],
-                    preFormattedNumber.split(" ")[2].replaceAll("\\D+", ""));
+
+            String[] connectionNumbers = Arrays.copyOfRange(splittedPreformattedNumber, 2,
+                    splittedPreformattedNumber.length);
+            String connectionNumber = StringUtils.join(connectionNumbers, "");
+
+            return new NationalNumber(splittedPreformattedNumber[1], connectionNumber.replaceAll("\\D+", ""));
         }
-
-        return areaCodeAndConnectionNumber;
     }
 
 
-    private String chunkAndFormatResult(int countryCode, AreaCodeAndConnectionNumber areaCodeAndConnectionNumber) {
+    private String chunkAndFormatResult(int countryCode, NationalNumber nationalNumber) {
 
-        final String numberFormat = "+%d %s %s";
+        logger().debug("chunk the number:'{} {} {}' into blocks of 4 digits. the last block can be 1-4 digits",
+            countryCode, nationalNumber.areaCode, nationalNumber.connectionNumber);
 
-        // chunk the number into blocks of 4 digits. the last block can be 1-4 digits
-        final String chunkedConnectionNumber = String.join(" ",
-                areaCodeAndConnectionNumber.connectionNumber.split("(?<=\\G.{4})"));
+        final String chunkedConnectionNumber = String.join(" ", nationalNumber.connectionNumber.split("(?<=\\G.{4})"));
 
-        // target format
-        return String.format(numberFormat, countryCode, areaCodeAndConnectionNumber.areaCode, chunkedConnectionNumber);
+        if (StringUtils.isBlank(nationalNumber.areaCode)) {
+            return String.format("+%d %s", countryCode, chunkedConnectionNumber);
+        } else {
+            return String.format("+%d %s %s", countryCode, nationalNumber.areaCode, chunkedConnectionNumber);
+        }
     }
 
-    private class AreaCodeAndConnectionNumber {
+
+    /**
+     * removes all whitespaces and line breaks.
+     *
+     * @param  phoneNumber  the number to be trimed
+     *
+     * @return  the trimed number
+     */
+    private String trimPhoneNumber(String phoneNumber) {
+
+        logger().debug("trim number:'{}' remove spaces, new Lines, '/', '-' and '()'.", phoneNumber);
+
+        String phoneNumberWithoutNewLine = phoneNumber.replaceAll("\\s+", "").replaceAll("\n", "");
+        String phoneNumberWithoutBraces = phoneNumberWithoutNewLine.replaceAll("\\(", "").replaceAll("\\)", "");
+
+        return phoneNumberWithoutBraces.replaceAll("/", "").replaceAll("-", "");
+    }
+
+    // National number contains areaCode and connectionNumber
+    private class NationalNumber {
 
         public final String areaCode;
         public final String connectionNumber;
 
-        private AreaCodeAndConnectionNumber(String areaCode, String connectionNumber) {
+        private NationalNumber(String areaCode, String connectionNumber) {
 
             this.areaCode = areaCode;
             this.connectionNumber = connectionNumber;